refactor: address coderabbit suggestions
