+++ web/package.json
@@ -46,7 +46,7 @@
     "@nuxt/eslint": "^1.0.0",
     "@nuxt/test-utils": "^3.17.2",
     "@nuxtjs/tailwindcss": "^6.12.2",
-    "@pinia/testing": "^0.1.0",
+    "@pinia/testing": "^1.0.0",
     "@rollup/plugin-strip": "^3.0.4",
     "@tailwindcss/typography": "^0.5.15",
     "@testing-library/vue": "^8.0.0",
